🏷️ Add `Table` type
The `Table` type is mainly used implicitly within the closures of many methods on readers and writers.
Its main use is to return `Column`s when indexed. The current leak-drop implementation is ok, but requires a bit more thought to make sure there are no disregarded ways to write unsafe code with it.