⚗️Naked Trait Objects + Trait Upcasting
This expression structure seems very promising, it solves the main two issues of expression coupling and implicit casting when dealing with operators.

A downside of this approach is that there is no implicit casting for literals. If these want to be used as the `self` types in functions they will need to go through an explicit cast (to be determined). The unstable nature of `trait_upcast` is not concerning as it has only been rolled back to unstable due to a conflict with another experimental feature. Within its use-cases here no such conflict occurs, and should therefore be safe and stable to use.