🚧🏷️ Add case expressions
WIP: `CoreExpr::conditions` is not fully implemented.

While working on the functionality of the case expression, a better implementation of conditional checking came to mind.
Iterators will be replaced by a concrete type that may contain one or more errors, as well as stack information or any other required data.
This is a much better interface than the current implementation.