Rework Types and Conditions
Types:

Expressions are now wrapped in a type (`CommonExpr` or `UniqueExpr`) to
differentiate between what previously were 'core' expressions and
'unique' expressions. This allows for a much simpler implementation of
the multi-expression functionality. Because of this the inner proc-macro
crate is no longer needed and has been removed.

Expressions no longer carry the type info of their children, i.e. before
one would see `Box<dyn Boolean>`. This was a nice handrail in the type
system when writing methods, ensuring that no mistakes were made, however
it meant that expression methods could not be added directly into the
`ExpresionType` (`Boolean`, `Textual`, ...) traits and instead required
new traits for each method in order to be implemented generically. This
added increasing amounts of bloat and was not deemed worth the type
safety tradeoff. Now all children are of the form
`Box<dyn Expression>`. All required methods have been moved to the
`Expression` trait leaving all the `ExpressionType` traits free to use
generics and implement all of the functionality of expressions.

Conditions:

Following the work on the 16-conditions branch, expressions are now
validated using a client-server protocol that passes the checker around
internally; validating expressions one by one instead of collecting all
of the conditions for an expression beforehand. This is a much cleaner
implementation allowing for the same freedom but removing all of the
heavy indirected iterators that existed.

Moving to the client-server protocol also allows for a cleaner API
between the `Builder`s/`Expr`s and the `Server`s/`Checker`s. For example
the current implementation makes the messages be of two types:
`Condition` or `Signal`. `Condition`s are meant for the standard
validations that the checker makes, while `Signal`s send other
information, such as context, a pre-checker error, or complex state
awareness such as expression coercion links in the case of a
`DerivedChecker`.